Refactor data fetching procedure out from api into common directory
